Add files via upload
In the  Python script, its start by importing the qrcode library, which simplifies the process of generating QR codes. The generate_qr_code function takes two parameters: data (the information to be encoded) and filename (the name of the output file where the QR code image will be saved).

Within the function, we create a QRCode instance, specifying parameters such as version, error correction level, box size, and border. We then add the data to be encoded using the add_data method and generate the QR code image using the make_image method.

Finally, the generated QR code image is saved using the save method with the specified filename.

To use this generator, simply call the generate_qr_code function with the data you want to encode and the desired output filename. In our example, we encoded a URL and saved the QR.